Fix & refactoring
Colors for filtering by "transferred" and "transferring" profiles

SDK queries centralized on src/data
